+++ app/api/wordcloud_update.py
@@ -1,6 +1,6 @@
 import pandas as pd
 from fastapi import APIRouter, Query
-from app.utils.wordcloud_functions import clean_text, complexity_df, count_syllables
+from app.utils.wordcloud.wordcloud_functions import clean_text, complexity_df, count_syllables
 
 router = APIRouter()
